Most of the battle enhancements

+++ Dragoon Modifier.Emulator/Memory/Battle/IBattle.cs
@@ -17,6 +17,6 @@ namespace Dragoon_Modifier.Emulator.Memory.Battle {
         uint MonsterPoint { get; }
         Monster[] MonsterTable { get; }
         ushort[] UniqueMonsterID { get; }
-        ushort ItemUsed { get; set; }
+        byte ItemUsed { get; set; }
     }
 }
